+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2020-10-04
+
+### Added
+- setTemplate method in Listing class, add support for custom templates
+
+### Changed
+- DataTablesContext, remove pathName used with old annotation solution
+
+### Removed
+- All deprecated methods
+- Annotations support for controllers
+- AbstractDataTablesController 
+
 ## [1.1.0] - 2020-04-25
 ### Added